test: split forecasting tests by db scope
